Remove Prototype-Patching in Repo-transaction
This should allow us to build the SDK without relying on module side-effects
